Fixed post processing chart bug from Mike's charts
There was a silent null pointer error introduced
that manifested if a prior run generated fills,
but the new run had fewer interests.  The new
run would try to generate charts for all
interests, pulling in extant fills, but finding
no deployments (from the new run).  This led
to an indirect, deeply buried null pointer
error when trying to synchronize all chart
scales and axes.

The fix is to pre-compute scoped interests
from deployments, then use that to constrain
which charts are actually built.  THis keeps
us away from empty data.  We also optimize
the build by re-using the pre-computed
deployment table and not trying to build
tons of charts.

Another possible fix would be to detect
absense of axes (i.e. an empty chart)
and just not synchronize the chart.

Problem here is we get tons of empty
charts now, and waste time building them.

Current approach is faster and leaner.