Extending sparkchart upperbound when necessary
Have to set the upper bound of the sparkchart at or above the
maximum of tadmudi and marathon % demand met.  When the upper bound
of the sparchart was being set to the max %, some of the tadmudi
line was being cut off so I extended the upper bound beyond that
a bit.