Added ability to compile run statistics
The analysts requested a comparison of two marathon runs and 1-3
run statistics in a table.  Added the fns required to do so in
a more general manner than what was done in proc.stats.
Now the caller can simply provide a sequence of statistics fns
which each operate on an audit trail dir location and a sequence of
the run directories.