Added namespace to ensure sparkcharts and tadmudi match
VBA TADMUDI results and the output from sparkchart functions should
match exactly for consistency in results provided to customer.
This wasn't the case at first, but after a few updates to TADMUDI,
the results for the committed test data match between both methods.
Functions are included to conduct the comparison.