Spark charts use supply policies now
Updated spark charts so that if a unique
policy is specified in SupplyRecords, then
that policy is used to compute the theoretical
capacity.  Added the latest version of
TADMUDI dataset to test resources in order
to test this functionality.  In TADMUDI,
the string representation of policies for
each period has changed so added that capability
here.