Periods and parameter from multiple sources
Before, they could only be loaded from tabular text files.  Now,
they can be loaded from a path to an Excel workbook or an
Excel workbook as a resource.

Also had to bump Spork in order to load a workbook from a
resource, which required the incanter-excel exclusion.

Came up with a more fool-proof way to find the last-day-default.

Committed now before I replace a bunch of functions with the
defrecord-getter function.